style according to figma
